refactor: refactor project entities, view mapper and transformer (#714)
