+++ explorer/app/components/Detail/components/GeneratedMatricesTables/components/FileLocationArchivePreview/fileLocationArchivePreview.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import { ProjectMatrixView } from "../../../../../../viewModelBuilders/azul/common/entities";
+import { ProjectMatrixView } from "../../../../../../viewModelBuilders/azul/hca-dcp/common/projectMatrixMapper/entities";
 import { ArchivePreviewDialog } from "../ArchivePreviewDialog/archivePreviewDialog";
 import { FileLocationArchivePreviewButton as Button } from "./fileLocationArchivePreview.styles";
