+++ explorer/app/components/Detail/components/GeneratedMatricesTables/components/FileLocationDownload/fileLocationDownload.tsx
@@ -3,7 +3,7 @@ import { ButtonGroupButton } from "@clevercanary/data-explorer-ui/lib/components
 import { DownloadIconSmall } from "@clevercanary/data-explorer-ui/lib/components/common/CustomIcon/common/constants";
 import { useFileLocation } from "@clevercanary/data-explorer-ui/lib/hooks/useFileLocation";
 import React from "react";
-import { ProjectMatrixView } from "../../../../../../viewModelBuilders/azul/common/entities";
+import { ProjectMatrixView } from "../../../../../../viewModelBuilders/azul/hca-dcp/common/projectMatrixMapper/entities";
 
 export interface FileLocationDownloadProps {
   projectMatrixView: ProjectMatrixView;
