+++ explorer/app/components/Detail/components/GeneratedMatricesTables/generatedMatricesTables.tsx
@@ -3,8 +3,8 @@ import { Table } from "@clevercanary/data-explorer-ui/lib/components/Detail/comp
 import { Divider } from "@mui/material";
 import { ColumnDef } from "@tanstack/react-table";
 import React, { Fragment } from "react";
-import { GenusSpecies } from "../../../../viewModelBuilders/azul/common/constants";
-import { ProjectMatrixTableView } from "../../../../viewModelBuilders/azul/common/entities";
+import { GenusSpecies } from "../../../../viewModelBuilders/azul/hca-dcp/common/projectMatrixMapper/constants";
+import { ProjectMatrixTableView } from "../../../../viewModelBuilders/azul/hca-dcp/common/projectMatrixMapper/entities";
 import { Alert, SectionTitle } from "./generatedMatricesTables.styles";
 
 interface GeneratedMatricesTablesProps<T extends object> {
