+++ explorer/app/hooks/useProjectMatrixArchiveFiles.ts
@@ -5,7 +5,7 @@ import {
   ArchiveFile,
   fetchProjectMatrixArchiveFiles,
 } from "../apis/azul/hca-dcp/common/archivePreview";
-import { ProjectMatrixView } from "../viewModelBuilders/azul/common/entities";
+import { ProjectMatrixView } from "../viewModelBuilders/azul/hca-dcp/common/projectMatrixMapper/entities";
 
 export interface UseArchivePreview {
   archiveFiles?: ArchiveFile[];
